pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "cnes-etl"
version = "0.1.0"
description = "CNES ETL (extract + transform)"
authors = [{name = "Caio Maximiano"}]
requires-python = ">=3.9"
dependencies = [
"pandas>=2.2",
"pyarrow>=16",
"azure-storage-file-datalake>=12.14",
"python-dateutil>=2.9",
"requests>=2.32",
"urllib3>=2",
"typer[all]>=0.12",
"tenacity>=8.2",
"truststore>=0.9",
"certifi>=2024.2",
]


[project.optional-dependencies]
dev = [
"pytest>=8",
"pytest-cov>=5",
"mypy>=1.10",
"ruff>=0.6",
"types-requests",
]


[tool.ruff]
line-length = 100
select = ["E","F","I","UP","B","SIM"]


[tool.mypy]
python_version = "3.9"
strict = true
warn_unused_ignores = true


[project.scripts]
cnes-etl = "cnes_etl.cli:app"